Updated unit tests

+++ core/cpp/infrastructure/srt_active_goals.cpp
@@ -16,7 +16,8 @@ void srt_active_goals::flush_srt_goal_batch() {
 }
 
 void srt_active_goals::erase_active_goal(const goal_lineage* gl) {
-    DEBUG_ASSERT(tree_.unlink(gl));
+    if (!tree_.roots().contains(gl))
+        DEBUG_ASSERT(tree_.unlink(gl));
     DEBUG_ASSERT(tree_.erase(gl));
 }
